requirements.txt
alembic==0.8.6
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.4.1
billiard==3.3.0.23
blinker==1.4
celery==3.1.23
Flask==0.10.1
Flask-Admin==1.4.0
Flask-AutoIndex==0.6
Flask-Failsafe==0.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-Silk==0.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
future==0.15.2
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
Mako==1.0.4
MarkupSafe==0.23
marshmallow==2.7.1
passlib==1.6.5
psycopg2==2.6.1
python-editor==1.0
pytz==2016.3
PyYAML==3.11
redis==2.10.5
requests==2.9.1
SQLAlchemy==1.0.12
Werkzeug==0.11.8
WTForms==2.1
